requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.17
chromadb>=0.4.22
sentence-transformers>=2.2.2
pyyaml>=6.0
ollama>=0.1.5
torch>=2.1.0
pdf2image>=1.16.3
pytesseract>=0.3.10
python-dotenv>=1.0.0 